IGNITE-28819 Move partition/safeAbs helpers to CommonUtils
Relocate the static helpers calculateMask(int) and calculatePartition(Object,
int, int) from RendezvousAffinityFunction, and safeAbs(int)/safeAbs(long) from
IgniteUtils, into ignite-commons CommonUtils, so the thin client
(ClientCacheAffinityMapping) can compute partitions without depending on
ignite-core. RendezvousAffinityFunction.calculateMask/calculatePartition now
delegate to CommonUtils; safeAbs keeps working at existing call sites via
IgniteUtils inheriting from CommonUtils.
