+++ modules/core/src/main/java/org/apache/ignite/internal/util/IgniteUtils.java
@@ -5552,30 +5552,6 @@ public abstract class IgniteUtils extends CommonUtils {
     }
 
     /**
-     * Gets absolute value for integer. If integer is {@link Integer#MIN_VALUE}, then {@code 0} is returned.
-     *
-     * @param i Integer.
-     * @return Absolute value.
-     */
-    public static int safeAbs(int i) {
-        i = Math.abs(i);
-
-        return i < 0 ? 0 : i;
-    }
-
-    /**
-     * Gets absolute value for long. If argument is {@link Long#MIN_VALUE}, then {@code 0} is returned.
-     *
-     * @param i Argument.
-     * @return Absolute value.
-     */
-    public static long safeAbs(long i) {
-        i = Math.abs(i);
-
-        return i < 0 ? 0 : i;
-    }
-
-    /**
      * When {@code long} value given is positive returns that value, otherwise returns provided default value.
      *
      * @param i Input value.
